backends: Silence type warning.

+++ libs/backends/GP_Backend.c
@@ -140,7 +140,7 @@ int gp_backend_timer_timeout(gp_backend *self)
 
 	uint64_t now = gp_time_stamp();
 
-	return GP_MAX(0, self->timers->expires - now);
+	return GP_MAX((uint64_t)0, self->timers->expires - now);
 }
 
 void gp_backend_poll(gp_backend *self)
